+++ NEWS.md
@@ -1,3 +1,7 @@
+# Changes in version 0.0.2
+
+* Updated package structure
+
 # Changes in version 0.0.1
 
-  * Added 2 demo shiny dashboards 
+* Added 2 demo shiny dashboards